Test and fix all routes

+++ server/src/routes/channel-router.ts
@@ -3,13 +3,16 @@ import { Router, Request, Response } from "express";
 import prisma from "../db/db";
 const router = Router();
 
-// /channel/:channelId -> For creating/editing that sub event 
-router.route("/:channelId")
-    .get(async (req: Request, res: Response) => {
+// /channel -> For creating/editing that sub event 
+router.route("/")
+    .post(async (req: Request, res: Response) => {
         try {
-            const { channelId } = req.params;
+            const { channelId, eventId } = req.body;
+            if (!channelId || !eventId) {
+                return res.status(401).json({ message: "Event id or Channel id missing!" });
+            }
             const channel = await prisma.channel.findUnique({
-                where: { id: Number(channelId) },
+                where: { id: Number(channelId), eventId: Number(eventId) },
                 include: {
                     ChannelService: true,
                     GroupRelation: true,
@@ -26,10 +29,14 @@ router.route("/:channelId")
     })
     .put(async (req: Request, res: Response) => {
         try {
-            const { channelId } = req.params;
-            const { name, venue, startTime, endTime } = req.body;
+            const { channelId, eventId, name, venue, startTime, endTime } = req.body;
+
+            if (!channelId || !eventId) {
+                return res.status(401).json({ message: "Event id or Channel id missing!" });
+            }
+
             const channel = await prisma.channel.update({
-                where: { id: Number(channelId) },
+                where: { id: Number(channelId), eventId: Number(eventId) },
                 data: { name, venue, startTime, endTime },
             });
             return res.status(201).json({ message: "Successfully changed the group details!", channel });
@@ -39,9 +46,14 @@ router.route("/:channelId")
     })
     .delete(async (req: Request, res: Response) => {
         try {
-            const { channelId } = req.params;
+            const { channelId, eventId } = req.body;
+
+            if (!channelId || !eventId) {
+                return res.status(401).json({ message: "Event id or Channel id missing!" });
+            }
+
             await prisma.channel.delete({
-                where: { id: Number(channelId) },
+                where: { id: Number(channelId), eventId: Number(eventId) },
             });
             res.status(204).json({ message: "Channel deleted successfully" });
         } catch (error) {
@@ -52,7 +64,8 @@ router.route("/:channelId")
 // /channel/group/create -> Creating a new group inside a sub event
 router.post("/group/create", async (req: Request, res: Response) => {
     try {
-        const { eventId, name, desc } = req.body;
+        const { eventId, channelId, name, desc } = req.body;
+
         const group = await prisma.group.create({
             data: {
                 name,
@@ -60,18 +73,38 @@ router.post("/group/create", async (req: Request, res: Response) => {
                 eventId,
             },
         });
-        res.json(group);
+
+        await prisma.groupRelation.create({
+            data: {
+                groupId: group.id,
+                channelId: channelId,
+            },
+        });
+
+        return res.status(201).json({ message: "Successfully created the group inside the channel!", group });
     } catch (error) {
         res.status(500).json({ error: "Internal server error" });
     }
 });
 
 // /channel/user -> Adding /removing a user from sub event
-router.post("/channel/user", async (req: Request, res: Response) => {
+// action -> "add" or "remove"
+router.post("/user", async (req: Request, res: Response) => {
     try {
         const { channelId, userId, action } = req.body;
 
         if (action === "add") {
+            const existingParticipant = await prisma.channelParticipant.findFirst({
+                where: {
+                    participantId: userId,
+                    channelId: channelId,
+                }
+            });
+
+            if (existingParticipant) {
+                return res.status(400).json({ message: "User already in the channel!" });
+            }
+
             const channelParticipant = await prisma.channelParticipant.create({
                 data: {
                     channelId,
@@ -86,26 +119,36 @@ router.post("/channel/user", async (req: Request, res: Response) => {
                     participantId: userId,
                 },
             });
-            return res.status(200).json({ message: "User removed from channel", channelParticipant });
+            return res.status(200).json({ message: "User removed from channel!", channelParticipant });
         } else {
             return res.status(400).json({ message: "Invalid action, try again!" });
         }
     } catch (error) {
+        console.log("Error in /channel/user", error);
         res.status(500).json({ error: "Internal server error" });
     }
 });
 
 // /channel/list -> Listing all users from that sub event
-router.get("/channel/list", async (req: Request, res: Response) => {
+router.get("/list", async (req: Request, res: Response) => {
     try {
         const { channelId } = req.query;
         const participants = await prisma.channelParticipant.findMany({
             where: { channelId: Number(channelId) },
             include: {
-                EventParticipant: true,
+                EventParticipant: {
+                    include: {
+                        User: true
+                    }
+                },
             },
         });
-        return res.status(200).json({ message: "Successfully fetched all the participants!", participants });
+        const users = [];
+        participants.map(participant => {
+            users.push({ ...participant?.EventParticipant?.User, "role": participant?.EventParticipant?.role })
+        });
+
+        return res.status(200).json({ message: "Successfully fetched all the participants!", "participants": users });
     } catch (error) {
         res.status(500).json({ error: "Internal server error" });
     }
